+++ src/main/java/vn/fpt/tranduykhanh/bookingservicepetshop/services/ServiceImplement.java
@@ -109,6 +109,10 @@ public class ServiceImplement implements ServiceInterface {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ResponseObj(HttpStatus.NOT_FOUND.toString(),"Not found", null));
         }
 
+        if(!existingServiceOpt.get().getBookingList().isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ResponseObj(HttpStatus.BAD_REQUEST.toString(),"Service này đang có booking", convertServiceToServiceResponseById(existingServiceOpt.get().getId())));
+        }
+
         if(serviceDTO.getServiceDescription() != null && !serviceDTO.getServiceDescription().isEmpty()){
             existingServiceOpt.get().setDescription(serviceDTO.getServiceDescription());
         }
@@ -163,6 +167,10 @@ public class ServiceImplement implements ServiceInterface {
 
         if (!serviceOpt.isPresent()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ResponseObj(HttpStatus.NOT_FOUND.toString(),"Service not found", null));
+        }
+
+        if(!serviceOpt.get().getBookingList().isEmpty()){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ResponseObj(HttpStatus.BAD_REQUEST.toString(),"Service này đang có booking", null));
         }
 
         serviceRepository.delete(serviceOpt.get());
